requirements.txt
appscript==1.2.2
async-generator==1.10
attrs==23.1.0
Babel==2.12.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.16.6
exceptiongroup==1.1.1
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
log21==2.5.1
lxml==4.9.2
MarkupSafe==2.1.2
msgpack==1.0.5
msoffice2pdf==0.0.8
numpy==1.24.3
opencv-python==4.7.0.72
outcome==1.2.0
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pypng==0.20220715.0
PySocks==1.7.1
python-docx==0.8.11
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.28.2
rsa==4.9
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
webcolors==1.13
wsproto==1.2.0
